# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Avaliação - Automação de Testes",
    "version": "1.0.0",
    "description": "API para autenticação, cadastro de usuários e gerenciamento de itens."
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registro de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["user", "admin"] }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "400": { "description": "Usuário já existe ou dados inválidos" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT válido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/items": {
      "get": {
        "summary": "Lista todos os itens",
        "responses": {
          "200": { "description": "Lista de itens" }
        }
      },
      "post": {
        "summary": "Cria novo usuário",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "role": { "type": "string", "enum": ["user", "admin"] }
                },
                "required": ["name", "role"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Não autenticado" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}